# Conflicts:
#	src/main/resources/application.properties

+++ src/main/java/com/example/finalproject/domain/menus/repository/MenuCategoriesRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface MenuCategoriesRepository extends JpaRepository<MenuCategories, Integer> {
+public interface MenuCategoriesRepository extends JpaRepository<MenuCategories, Long> {
     List<MenuCategories> findByMenuId(Long menuId);
 }
